# brain-backend/api-ref/brain_openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Brain API",
        "version": "0.1.0"
    },
    "paths": {
...
        "/system-disks": {
            "get": {
                "tags": [
                    "block"
                ],
                "summary": "Get All System Disks",
                "description": "Get all system disks list",
                "operationId": "get_all_system_disks_system_disks_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/SystemDisk"
                                    },
                                    "type": "array",
                                    "title": "Response Get All System Disks System Disks Get"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "block"
                ],
                "summary": "Create System Disk",
                "description": "Create a new system disk RBD from image for specific MV200 server",
                "operationId": "create_system_disk_system_disks_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BareMetalCreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SystemDisk"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/system-disks/{disk_id}": {
            "get": {
                "tags": [
                    "block"
                ],
                "summary": "Get System Disk",
                "description": "Get specific system disk by ID",
                "operationId": "get_system_disk_system_disks__disk_id__get",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Disk Id"
                        },
                        "name": "disk_id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SystemDisk"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "block"
                ],
                "summary": "Update System Disk",
                "description": "Update system disk information by ID",
                "operationId": "update_system_disk_system_disks__disk_id__put",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Disk Id"
                        },
                        "name": "disk_id",
                        "in": "path"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SystemDiskUpdate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SystemDisk"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "block"
                ],
                "summary": "Delete System Disk",
                "description": "Delete system disk by ID",
                "operationId": "delete_system_disk_system_disks__disk_id__delete",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Disk Id"
                        },
                        "name": "disk_id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "BareMetalCreate": {
                "properties": {
                    "system_disk": {
                        "$ref": "#/components/schemas/SystemDiskCreate"
                    },
                    "system_user": {
                        "$ref": "#/components/schemas/SystemUser"
                    }
                },
                "type": "object",
                "required": [
                    "system_disk",
                    "system_user"
                ],
                "title": "BareMetalCreate"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "SystemDisk": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id",
                        "description": "Unique identifier for the system disk"
                    },
                    "image_id": {
                        "type": "string",
                        "title": "Image Id",
                        "description": "Source image id"
                    },
                    "mv200_id": {
                        "type": "string",
                        "title": "Mv200 Id",
                        "description": "Target MV200 server id"
                    },
                    "mv200_ip": {
                        "type": "string",
                        "title": "Mv200 Ip",
                        "description": "Target MV200 server IP address"
                    },
                    "size_gb": {
                        "type": "integer",
                        "minimum": 1,
                        "title": "Size Gb",
                        "description": "Disk size in GB"
                    },
                    "mon_host": {
                        "type": "string",
                        "format": "ipvanyaddress",
                        "title": "Mon Host",
                        "description": "Ceph monitor host IP address"
                    },
                    "rbd_path": {
                        "type": "string",
                        "title": "Rbd Path",
                        "description": "Target rbd in ceph cluster"
                    },
                    "blk_id": {
                        "type": "integer",
                        "title": "Blk Id",
                        "description": "Target blk uuid in soc"
                    },
                    "description": {
                        "type": "string",
                        "title": "Description",
                        "description": "Disk description"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "image_id",
                    "mv200_id",
                    "mv200_ip",
                    "size_gb",
                    "mon_host",
                    "rbd_path",
                    "blk_id"
                ],
                "title": "SystemDisk"
            },
            "SystemDiskCreate": {
                "properties": {
                    "image_id": {
                        "type": "string",
                        "title": "Image Id",
                        "description": "Source image id"
                    },
                    "mv200_id": {
                        "type": "string",
                        "title": "Mv200 Id",
                        "description": "Target MV200 server id"
                    },
                    "size_gb": {
                        "type": "integer",
                        "minimum": 1,
                        "title": "Size Gb",
                        "description": "Disk size in GB"
                    },
                    "description": {
                        "type": "string",
                        "title": "Description",
                        "description": "Disk description"
                    }
                },
                "type": "object",
                "required": [
                    "image_id",
                    "mv200_id",
                    "size_gb"
                ],
                "title": "SystemDiskCreate"
            },
            "SystemDiskUpdate": {
                "properties": {
                    "description": {
                        "type": "string",
                        "title": "Description",
                        "description": "Disk description"
                    }
                },
                "type": "object",
                "title": "SystemDiskUpdate"
            },
            "SystemUser": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name",
                        "description": "System user name"
                    },
                    "password": {
                        "type": "string",
                        "title": "Password",
                        "description": "System user password"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "password"
                ],
                "title": "SystemUser"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "login"
                    }
                }
            }
        }
    }
}